requirements.txt
Pillow
python-dotenv
requests
cachetools
wmi
pyspectator